task states and logs do not inherit generated models so there will not be duplicates
